sst_backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
requests==2.32.5
beautifulsoup4==4.14.3
python-dotenv==1.2.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
dj-database-url==3.1.2
